+++ components/user-mgt/org.wso2.carbon.user.mgt/src/main/java/org/wso2/carbon/user/mgt/RolePermissionManagementService.java
@@ -17,7 +17,7 @@
  */
 package org.wso2.carbon.user.mgt;
 
-import org.wso2.carbon.user.mgt.common.Permission;
+import org.wso2.carbon.user.mgt.common.model.Permission;
 
 /**
  * OSGi service interface which use to manage role permissions.
